refactor: 테마 디자인 구조와 iOS 표시를 정리
