+++ client/iosApp/MapmoryApp.swift
@@ -1,40 +1,76 @@
 import Shared
 import SwiftUI
 
-private let systemBarColor = Color(
+private let lightSystemBarColor = UIColor(
     red: 250.0 / 255.0,
     green: 252.0 / 255.0,
-    blue: 251.0 / 255.0
+    blue: 251.0 / 255.0,
+    alpha: 1
+)
+private let darkSystemBarColor = UIColor(
+    red: 17.0 / 255.0,
+    green: 21.0 / 255.0,
+    blue: 24.0 / 255.0,
+    alpha: 1
 )
 
 @main
 struct MapmoryApp: App {
+    @State private var isDarkTheme = false
+
     var body: some Scene {
         WindowGroup {
             ZStack {
-                systemBarColor
-                ComposeView()
+                Color(uiColor: isDarkTheme ? darkSystemBarColor : lightSystemBarColor)
+                ComposeView(isDarkTheme: $isDarkTheme)
             }
             .ignoresSafeArea()
-            .preferredColorScheme(.light)
+            .preferredColorScheme(isDarkTheme ? .dark : .light)
         }
     }
 }
 
 private struct ComposeView: UIViewControllerRepresentable {
+    @Binding var isDarkTheme: Bool
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(isDarkTheme: $isDarkTheme)
+    }
+
     func makeUIViewController(context: Context) -> UIViewController {
-        let viewController = MainViewControllerKt.MainViewController()
-        viewController.view.backgroundColor = UIColor(
-            red: 250.0 / 255.0,
-            green: 252.0 / 255.0,
-            blue: 251.0 / 255.0,
-            alpha: 1,
-        )
+        let coordinator = context.coordinator
+        let viewController = MainViewControllerKt.MainViewController { isDark in
+            coordinator.updateTheme(isDark.boolValue)
+        }
+        applyTheme(to: viewController)
         return viewController
     }
 
     func updateUIViewController(
         _ uiViewController: UIViewController,
         context: Context
-    ) {}
+    ) {
+        context.coordinator.isDarkTheme = $isDarkTheme
+        applyTheme(to: uiViewController)
+    }
+
+    private func applyTheme(to viewController: UIViewController) {
+        viewController.overrideUserInterfaceStyle = isDarkTheme ? .dark : .light
+        viewController.view.backgroundColor = isDarkTheme ? darkSystemBarColor : lightSystemBarColor
+        viewController.setNeedsStatusBarAppearanceUpdate()
+    }
+
+    final class Coordinator {
+        var isDarkTheme: Binding<Bool>
+
+        init(isDarkTheme: Binding<Bool>) {
+            self.isDarkTheme = isDarkTheme
+        }
+
+        func updateTheme(_ isDark: Bool) {
+            DispatchQueue.main.async { [weak self] in
+                self?.isDarkTheme.wrappedValue = isDark
+            }
+        }
+    }
 }
