+++ client/shared/src/iosMain/kotlin/com/mapmory/shared/MainViewController.kt
@@ -4,6 +4,9 @@ import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.safeDrawing
 import androidx.compose.ui.window.ComposeUIViewController
 
-fun MainViewController() = ComposeUIViewController {
-    MapmoryApp(contentWindowInsets = WindowInsets.safeDrawing)
+fun MainViewController(onThemeChanged: (Boolean) -> Unit) = ComposeUIViewController {
+    MapmoryApp(
+        contentWindowInsets = WindowInsets.safeDrawing,
+        onThemeChanged = onThemeChanged,
+    )
 }
